로그인하지 않아도 설문에 참여할 수 있습니다

+++ app/controllers/feedbacks_controller.rb
@@ -1,25 +1,53 @@
 class FeedbacksController < ApplicationController
-  before_filter :authenticate_user!
+  include FeedbackHelper
 
   def create
     @option = Option.find params[:option_id]
     survey = @option.survey
 
     if survey.open?
-      feedbacks = survey.feedbacks.where user: current_user
-      previous_feedback = feedbacks.find_by(option: @option)
-      is_present_previous_feedback = previous_feedback.present?
+      ActiveRecord::Base.transaction do
+        if user_signed_in?
+          feedbacks = survey.feedbacks.where user: current_user
+          previous_feedback = feedbacks.find_by(option: @option)
+          is_present_previous_feedback = previous_feedback.present?
 
-      if survey.multi_selectable?
-        if is_present_previous_feedback
-          previous_feedback.destroy
+          if survey.multi_selectable?
+            if is_present_previous_feedback
+              previous_feedback.destroy
+            else
+              feedback = create_feedback(@option)
+            end
+          else
+            feedbacks.destroy_all
+            unless is_present_previous_feedback
+              feedback = create_feedback(@option)
+            end
+          end
         else
-          feedback = create_feedback(@option)
-        end
-      else
-        feedbacks.destroy_all
-        unless is_present_previous_feedback
-          feedback = create_feedback(@option)
+          unless @option.survey.anonymous_feedbackable?
+            render 'login_alert.js.erb'
+            return
+          end
+
+          @previous_selected_options = @option.survey.options.where(id: fetch_anonymous_selected_option_ids(@option.survey))
+          if survey.multi_selectable?
+            if @previous_selected_options.include? @option
+              @option.decrement(:anonymous_feedbacks_count)
+              mark_anonymous_unselected_option(@option.survey, @option)
+            else
+              @option.increment(:anonymous_feedbacks_count)
+              mark_anonymous_selected_option(@option)
+            end
+          else
+            @option.survey.options.update_all(anonymous_feedbacks_count: 0)
+            mark_anonymous_unselected_option(@option.survey)
+            if @previous_selected_options.empty? or !@previous_selected_options.include? @option
+              @option.increment(:anonymous_feedbacks_count)
+              mark_anonymous_selected_option(@option)
+            end
+          end
+          @option.save
         end
       end
     end
